for_docker/requirements.txt
Flask_RESTful==0.3.9
Flask==1.1.2
selenium==3.141.0
webdriver_manager==3.4.2
requests==2.25.1
pandas==1.0.3
Pillow==8.3.1
